# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /medicines:
    get:
      operationId: getMedicines
      tags: [medicines]
      summary: List medicines
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: filter
          in: query
          schema:
            type: string
            enum: [all, low_stock, expired, near_expiry]
        - name: category_id
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of medicines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"
    post:
      operationId: createMedicine
      tags: [medicines]
      summary: Create a new medicine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
  /medicines/{id}:
    get:
      operationId: getMedicine
      tags: [medicines]
      summary: Get medicine by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    patch:
      operationId: updateMedicine
      tags: [medicines]
      summary: Update medicine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicineUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
    delete:
      operationId: deleteMedicine
      tags: [medicines]
      summary: Delete medicine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /medicines/barcode/{barcode}:
    get:
      operationId: getMedicineByBarcode
      tags: [medicines]
      summary: Get medicine by barcode
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Medicine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
        "404":
          description: Not found
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: List categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: List expenses
      parameters:
        - name: date_from
          in: query
          schema:
            type: ["string", "null"]
        - name: date_to
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Medicine:
      type: object
      required: [id, name, quantity, salePrice, purchasePrice]
      properties:
        id:
          type: integer
        barcode:
          type: ["string", "null"]
        name:
          type: string
        scientificName:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        purchasePrice:
          type: number
        salePrice:
          type: number
        quantity:
          type: integer
        productionDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        reorderLevel:
          type: ["integer", "null"]
        createdAt:
          type: string
    MedicineInput:
      type: object
      required: [name, salePrice, purchasePrice, quantity]
      properties:
        barcode:
          type: ["string", "null"]
        name:
          type: string
        scientificName:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        purchasePrice:
          type: number
        salePrice:
          type: number
        quantity:
          type: integer
        productionDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        reorderLevel:
          type: ["integer", "null"]
    MedicineUpdate:
      type: object
      properties:
        barcode:
          type: ["string", "null"]
        name:
          type: ["string", "null"]
        scientificName:
          type: ["string", "null"]
        manufacturer:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        location:
          type: ["string", "null"]
        purchasePrice:
          type: ["number", "null"]
        salePrice:
          type: ["number", "null"]
        quantity:
          type: ["integer", "null"]
        productionDate:
          type: ["string", "null"]
        expiryDate:
          type: ["string", "null"]
        reorderLevel:
          type: ["integer", "null"]
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: ["string", "null"]
        medicineCount:
          type: integer
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        icon:
          type: ["string", "null"]
...
    Customer:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        balance:
          type: number
        notes:
          type: ["string", "null"]
        totalSales:
          type: number
    CustomerInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        balance:
          type: number
        notes:
          type: ["string", "null"]
...
    Expense:
      type: object
      required: [id, category, amount, createdAt]
      properties:
        id:
          type: integer
        category:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        createdAt:
          type: string
    ExpenseInput:
      type: object
      required: [category, amount]
      properties:
        category:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
...